start styling
