--- src/components/banner inpage/banner-inpage-print.scss
@@ -1,239 +1,0 @@
-/*
- * Banner Inpage print
- * @define banner-inpage
- */
-
-@use 'sass:map';
-
-// Exposed variables
-$theme: null !default;
-$banner-inpage: null !default;
-$banner-inpage-print: null !default;
-
-.ecl-banner {
-  --banner-footer-height: 0px;
-
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-bottom: 0;
-  margin-inline-end: 0;
-  margin-top: 0;
-  padding-bottom: var(--banner-footer-height);
-  position: relative;
-}
-
-.ecl-banner__picture-container {
-  height: 100%;
-  margin: 0;
-  position: static;
-  width: 100%;
-}
-
-.ecl-banner__video-container {
-  height: 100%;
-  object-fit: cover;
-  overflow: hidden;
-  position: static;
-  width: 100%;
-}
-
-.ecl-banner__image {
-  display: block;
-  height: 100%;
-  object-fit: cover;
-  object-position: top;
-  position: relative;
-  width: 100%;
-}
-
-.ecl-banner__video {
-  display: block;
-  height: 100%;
-  object-fit: cover;
-  object-position: center;
-  position: relative;
-  width: 100%;
-}
-
-.ecl-banner .ecl-container {
-  padding-inline: map.get($banner-inpage-print, 'container-padding-horizontal');
-  position: absolute;
-  z-index: 1;
-}
-
-.ecl-banner__container {
-  display: flex;
-  padding-bottom: map.get($banner-inpage-print, 'container-padding-bottom');
-  padding-top: map.get($banner-inpage-print, 'container-padding-top');
-  width: 100%;
-
-  .ecl-banner--v-bottom & {
-    padding-bottom: calc(
-      #{map.get($banner-inpage-print, 'container-padding-bottom')} +
-        var(--banner-footer-height)
-    );
-  }
-}
-
-.ecl-banner__content {
-  align-items: flex-start;
-  background-color: map.get($banner-inpage, 'content-background');
-  border-radius: map.get($banner-inpage, 'content-border-radius');
-  box-sizing: content-box;
-  display: inline-flex;
-  flex-direction: column;
-  max-width: map.get($banner-inpage-print, 'content-width');
-  padding: map.get($banner-inpage-print, 'content-padding');
-
-  .ecl-banner--box-bg-none & {
-    background-color: transparent;
-    border-radius: 0;
-    padding: 0;
-  }
-
-  .ecl-banner--box-bg-dark & {
-    box-decoration-break: clone;
-    background: var(--cm-surface-transparent-70, var(--c-ov-l));
-    backdrop-filter: blur(2px);
-  }
-}
-
-.ecl-banner__content > *:first-child {
-  margin-top: 0;
-}
-
-.ecl-banner__title,
-.ecl-banner__title .ecl-banner__title-link {
-  color: map.get($banner-inpage, 'text-color', 'dark');
-  display: inline;
-  font: map.get($banner-inpage-print, 'font', 'title', 'desktop-m');
-  font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
-  margin: 0;
-
-  .ecl-banner--font-l & {
-    font: map.get($banner-inpage-print, 'font', 'title', 'desktop-l');
-    font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
-  }
-
-  .ecl-banner--color-light & {
-    color: map.get($banner-inpage, 'text-color', 'light');
-  }
-
-  &:visited {
-    color: map.get($banner-inpage, 'text-color', 'dark');
-  }
-}
-
-.ecl-banner__description,
-.ecl-banner__description .ecl-banner__description-link {
-  color: map.get($banner-inpage, 'text-color', 'dark');
-  font: map.get($banner-inpage-print, 'font', 'description', 'desktop-m');
-  margin-bottom: 0;
-  margin-top: map.get($theme, 'spacing-print', 's');
-
-  .ecl-banner--font-l & {
-    font: map.get($banner-inpage-print, 'font', 'description', 'desktop-l');
-  }
-
-  .ecl-banner--color-light & {
-    color: map.get($banner-inpage, 'text-color', 'light');
-  }
-
-  &:visited {
-    color: map.get($banner-inpage, 'text-color', 'dark');
-  }
-}
-
-.ecl-banner__credit {
-  background-color: map.get($banner-inpage, 'background-color');
-  color: var(--cm-on-surface-brand, var(--c-d));
-  font: map.get($banner-inpage-print, 'font', 'credit');
-  padding-bottom: map.get($theme, 'spacing-print', '2xs');
-  padding-top: map.get($theme, 'spacing-print', '2xs');
-  position: absolute;
-  width: 100%;
-  z-index: 1;
-
-  .ecl-banner & .ecl-container {
-    position: relative;
-  }
-}
-
-// No full width banner on print
-.ecl-banner--full-width {
-  border-radius: 0;
-  left: 50%;
-  margin-inline-start: -50vw;
-  width: 100vw;
-}
-
-// Medium banner - ratio 4:1
-.ecl-banner,
-.ecl-banner--m {
-  .ecl-banner__image,
-  .ecl-banner__video {
-    aspect-ratio: 4/1;
-
-    --css-aspect-ratio: 4/1;
-  }
-}
-
-// Extra small banner - ratio 6:1
-.ecl-banner--xs .ecl-banner__image,
-.ecl-banner--xs .ecl-banner__video {
-  aspect-ratio: 6/1;
-
-  --css-aspect-ratio: 6/1;
-}
-
-// Small banner - ratio 5:1
-.ecl-banner--s .ecl-banner__image,
-.ecl-banner--s .ecl-banner__video {
-  aspect-ratio: 5/1;
-
-  --css-aspect-ratio: 5/1;
-}
-
-// Large banner - ratio 3:1
-.ecl-banner--l .ecl-banner__image,
-.ecl-banner--l .ecl-banner__video {
-  aspect-ratio: 3/1;
-
-  --css-aspect-ratio: 3/1;
-}
-
-// Horizontal alignment
-.ecl-banner--h-center {
-  .ecl-banner__container {
-    justify-content: center;
-  }
-
-  .ecl-banner__content {
-    align-items: center;
-    text-align: center;
-  }
-}
-
-.ecl-banner--h-right {
-  .ecl-banner__container {
-    justify-content: flex-end;
-  }
-
-  .ecl-banner__content {
-    align-items: flex-end;
-    text-align: right;
-  }
-}
-
-// Vertical alignment
-// stylelint-disable-next-line no-descending-specificity
-.ecl-banner--v-top .ecl-container {
-  top: 0;
-}
-
-// stylelint-disable-next-line no-descending-specificity
-.ecl-banner--v-bottom .ecl-container {
-  bottom: 0;
-}
