+++ src/components/banner-inpage/.npmignore
