+++ src/components/banner-inpage/banner-inpage.scss
@@ -1,0 +1,323 @@
+/**
+ * Banner inpage
+ * @define banner-inpage
+ */
+
+@use 'sass:map';
+@use '@ecl/grid/mixins/breakpoints';
+
+// Exposed variables
+$theme: null !default;
+$banner-inpage: null !default;
+
+.ecl-banner-inpage {
+  align-items: center;
+  border-radius: map.get($banner-inpage, 'border-radius', 'banner');
+  box-shadow: map.get($banner-inpage, 'box-shadow');
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin-bottom: 0;
+  margin-inline-end: 0;
+  margin-top: 0;
+  position: relative;
+}
+
+.ecl-banner-inpage__picture-container {
+  margin: 0;
+  position: relative;
+  width: 100%;
+}
+
+.ecl-banner-inpage__video-container {
+  object-fit: cover;
+  overflow: hidden;
+  position: relative;
+  width: 100%;
+}
+
+.ecl-banner-inpage__container {
+  background-color: red;
+  bottom: 0;
+  height: 241px; // var
+  position: absolute;
+  width: 100%;
+}
+
+.ecl-banner-inpage__credit {
+  color: var(--cm-on-surface-brand, var(--c-d));
+  font: map.get($banner-inpage, 'font', 'credit', 'mobile');
+  margin-block-end: calc(-1 * var(--s-xl));
+  padding-block-start: 1px;
+  position: relative;
+  text-align: end;
+  width: 100%;
+  z-index: 1;
+}
+
+.ecl-banner-inpage__image {
+  display: block;
+  height: 100%;
+  max-height: 100%;
+  object-fit: cover;
+  object-position: top;
+  position: relative;
+  width: 100%;
+}
+
+.ecl-banner-inpage__video {
+  display: block;
+  height: 100%;
+  max-height: 100%;
+  object-fit: cover;
+  object-position: center;
+  position: relative;
+  width: 100%;
+}
+
+.ecl-banner-inpage__title,
+.ecl-banner-inpage__title-link {
+  color: #fff; // var
+  display: block;
+  font: map.get($banner-inpage, 'font', 'title', 'mobile-m');
+  font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+  line-height: map.get($banner-inpage, 'font', 'title', 'line-height-mobile-m');
+  padding: var(--s-xl) var(--s-l); // var
+
+  &:hover,
+  &:visited {
+    color: #fff; // var
+  }
+
+  .ecl-banner-inpage--font-s & {
+    font: map.get($banner-inpage, 'font', 'title', 'mobile-s');
+    font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+    line-height: map.get(
+      $banner-inpage,
+      'font',
+      'title',
+      'line-height-mobile-s'
+    );
+  }
+
+  .ecl-banner-inpage--font-l & {
+    font: map.get($banner-inpage, 'font', 'title', 'mobile-l');
+    font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+    line-height: map.get(
+      $banner-inpage,
+      'font',
+      'title',
+      'line-height-mobile-l'
+    );
+  }
+}
+
+.ecl-banner-inpage__play.ecl-button--icon-only,
+.ecl-banner-inpage__pause.ecl-button--icon-only {
+  align-items: center;
+  background-color: var(--cm-on-surface-brand, var(--c-d));
+  border: none;
+  border-radius: 50%;
+  bottom: var(--s-m);
+  color: #fff;
+  justify-content: center;
+  min-height: 32px;
+  min-width: 32px;
+  outline: #fff;
+  outline-offset: -4px;
+  position: absolute;
+  right: var(--s-m);
+  z-index: 2;
+
+  &:hover,
+  &:active,
+  &:focus-visible {
+    background-color: #fff;
+    border: none;
+    color: var(--cm-on-surface-brand, var(--c-d));
+    outline-color: var(--cm-on-surface-brand, var(--c-d));
+    padding: 0;
+  }
+
+  .ecl-button__container {
+    height: 32px;
+    width: 32px;
+  }
+
+  .ecl-button__icon {
+    height: 100%;
+    width: 100%;
+  }
+}
+
+// Full width
+.ecl-banner-inpage--full-width {
+  border-radius: 0;
+  left: 50%;
+  margin-inline-start: -50vw;
+  width: 100vw;
+}
+
+// Fixed ratio on mobile
+.ecl-banner-inpage__picture-container,
+.ecl-banner-inpage__video-container,
+.ecl-banner-inpage .ecl-banner-inpage__image,
+.ecl-banner-inpage .ecl-banner-inpage__video {
+  aspect-ratio: 3/2;
+
+  --css-aspect-ratio: 3/2;
+}
+
+/* stylelint-disable-next-line order/order */
+@include breakpoints.up('m') {
+  .ecl-banner-inpage .ecl-container {
+    padding-inline: map.get(
+      $banner-inpage,
+      'container-padding-inline',
+      'tablet'
+    );
+  }
+
+  // Fixed ratio on tablet
+  .ecl-banner-inpage__picture-container,
+  .ecl-banner-inpage__video-container,
+  .ecl-banner-inpage .ecl-banner-inpage__image,
+  .ecl-banner-inpage .ecl-banner-inpage__video {
+    aspect-ratio: 3/1;
+
+    --css-aspect-ratio: 3/1;
+  }
+
+  .ecl-banner-inpage__title,
+  .ecl-banner-inpage__title-link {
+    font: map.get($banner-inpage, 'font', 'title', 'tablet-m');
+    font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+    line-height: map.get(
+      $banner-inpage,
+      'font',
+      'title',
+      'line-height-tablet-m'
+    );
+
+    .ecl-banner-inpage--font-s & {
+      font: map.get($banner-inpage, 'font', 'title', 'tablet-s');
+      font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+      line-height: map.get(
+        $banner-inpage,
+        'font',
+        'title',
+        'line-height-tablet-s'
+      );
+    }
+
+    .ecl-banner-inpage--font-l & {
+      font: map.get($banner-inpage, 'font', 'title', 'tablet-l');
+      font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+      line-height: map.get(
+        $banner-inpage,
+        'font',
+        'title',
+        'line-height-tablet-l'
+      );
+    }
+  }
+}
+
+/* stylelint-disable-next-line order/order */
+@include breakpoints.up('xl') {
+  .ecl-banner-inpage {
+    aspect-ratio: 4/1;
+
+    --css-aspect-ratio: 4/1;
+  }
+
+  .ecl-banner-inpage .ecl-container {
+    padding-inline: map.get(
+      $banner-inpage,
+      'container-padding-inline',
+      'desktop'
+    );
+    position: absolute;
+    z-index: 1;
+  }
+
+  .ecl-banner-inpage__credit {
+    font: map.get($banner-inpage, 'font', 'credit', 'desktop');
+    margin-block-end: 0;
+    padding-block: var(--s-2xs);
+    position: absolute;
+    text-align: start;
+
+    .ecl-banner-inpage & .ecl-container {
+      position: relative;
+    }
+  }
+
+  // stylelint-disable-next-line plugin/selector-bem-pattern
+  .ecl-container
+    .ecl-banner-inpage:not(.ecl-banner-inpage--full-width)
+    .ecl-container {
+    padding-inline-start: var(--s-2xl);
+    padding-inline-end: var(--s-2xl);
+  }
+
+  .ecl-banner-inpage__play.ecl-button--icon-only,
+  .ecl-banner-inpage__pause.ecl-button--icon-only {
+    min-height: 40px;
+    min-width: 40px;
+
+    .ecl-button__container {
+      height: 40px;
+      width: 40px;
+    }
+  }
+
+  .ecl-banner-inpage__title,
+  .ecl-banner-inpage__title-link {
+    display: inline;
+    font: map.get($banner-inpage, 'font', 'title', 'desktop-m');
+    font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+    line-height: map.get(
+      $banner-inpage,
+      'font',
+      'title',
+      'line-height-desktop-m'
+    );
+    width: auto;
+
+    .ecl-banner-inpage--font-s & {
+      font: map.get($banner-inpage, 'font', 'title', 'desktop-s');
+      font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+      line-height: map.get(
+        $banner-inpage,
+        'font',
+        'title',
+        'line-height-desktop-s'
+      );
+    }
+
+    .ecl-banner-inpage--font-l & {
+      font: map.get($banner-inpage, 'font', 'title', 'desktop-l');
+      font-weight: map.get($banner-inpage, 'font', 'title', 'font-weight');
+      line-height: map.get(
+        $banner-inpage,
+        'font',
+        'title',
+        'line-height-desktop-l'
+      );
+    }
+  }
+
+  .ecl-banner-inpage__picture-container,
+  .ecl-banner-inpage__video-container {
+    aspect-ratio: auto;
+    height: 100%;
+    position: static;
+  }
+
+  .ecl-banner-inpage__picture {
+    display: block;
+    height: 100%;
+    position: relative;
+  }
+}
