+++ src/components/banner-inpage/demo/data--image.js
