+++ src/components/banner-inpage/package.json
